[skip ci] Build and base images for executions
